pyproject.toml
[project]
name = "ai-agents-gl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.177.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.7",
    "mermaid-python>=0.1",
    "openai>=1.106.1",
    "python-dotenv>=1.1.1",
    "yfinance>=0.2.65",
]
